requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
jinja2>=3.1.0
python-multipart>=0.0.6
anthropic>=0.40.0
apscheduler>=3.10.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
